moved the byte order logic to the core bus interface; #include cleanup
